Refactor SpotlessExtension into a legacy and non-legacy portion.
